Move docker image initialization out to it's own package.
